backend/pyproject.toml
[project]
name = "salesops-backend"
version = "0.1.0"
description = "SalesOps AI backend"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.109",
  "uvicorn[standard]>=0.30",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "aiosqlite>=0.20",
  "alembic>=1.13",
  "pydantic-settings>=2.2",
  "email-validator>=2.2",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "python-multipart>=0.0.9",
  "httpx>=0.27",
  "redis>=5.0",
  "celery>=5.4",
  "pgvector>=0.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
